A* final hopefully

+++ src/planning/astar.hpp
@@ -41,7 +41,7 @@ struct Node
         // std::cout << "fuck" << std::endl;
         // std::cout << "distance  : "  << map(cell.x, cell.y) << " " << mindist << std::endl;
         if(map(cell.x, cell.y) == -1) return false;
-        return map(cell.x, cell.y)  <= mindist;
+        return map(cell.x, cell.y)  <= (float)mindist;
         // return false;
     }
